Add mysqli extension and email check functionality
The mysqli PHP extension has been added to the 'require' key in the 'composer.json', as well as to the 'platform' key in the 'composer.lock' file. Additionally, two PHP scripts have been created, 'p.php' and 'check_email.php', for DNS checks on email addresses. This assists in validating user domains and handling instances of invalid email domains.